requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-apscheduler==0.5.1
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
pathspec==0.8.1
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
